extracted file manager and record manager logic from main to separate files
